requirements.txt
# --- Core Application & Web Server ---
fastapi
uvicorn[standard]
python-multipart

# --- GraphQL ---
strawberry-graphql[fastapi]

# --- Database & Migrations ---
sqlalchemy
alembic
asyncpg
psycopg2-binary

# --- Asynchronous Task & Caching ---
celery
redis
aioredis

# --- Authentication & Security ---
python-jose[cryptography]
passlib[bcrypt]

# --- NLP & Machine Learning ---
torch
transformers
spacy

# --- Data Handling & Web Scraping ---
pandas
openpyxl
beautifulsoup4
requests
newsapi-python

# --- Scheduling & Utilities ---
apscheduler
python-dotenv
httpx
pydantic
pydantic-settings

# --- Testing ---
pytest
pytest-asyncio